ai-module/requirements.txt
flask==3.1.0
opencv-python==4.11.0.86
python-dotenv==1.1.0
requests==2.32.3
ultralytics==8.3.125
